refactor: update data types for pricing in Bill and Order responses; adjust related queries and endpoints

+++ RestaurantManagement.Application/Features/BillFeature/Queries/GetAllBill/GetAllBillQueryHandler.cs
@@ -78,16 +78,16 @@ public class GetAllBillQueryHandler(
                 a.Booking.BookingDate,
                 a.Booking.BookingTime,
                 a.Order.OrderId,
-                a.Total,
+                (int)a.Total,
                 a.PaymentType,
                 a.Order.OrderDetails.Select(b => new OrderDetailResponse(
                     b.OrderDetailId,
                     b.MealId,
                     b.Meal.MealName,
-                    b.Meal.Price,
+                    (int)b.Meal.Price,
                     b.Meal.ImageUrl,
                     b.Quantity,
-                    b.UnitPrice
+                    (int)b.UnitPrice
                 )).ToArray())).AsQueryable();
 
         var billsList = await PagedList<BillResponse>.CreateAsync(bills, request.page ?? 1, request.pageSize ?? 10);
