+++ RestaurantManagement.Application/Features/OrderFeature/Queries/GetAllOrder/GetAllOrderQueryHandler.cs
@@ -72,15 +72,15 @@ public class GetAllOrderQueryHandler(
                 a.TableId,
                 a.OrderId,
                 a.PaymentStatus,
-                a.Total,
+                (int)a.Total,
                 a.OrderDetails.Select(b => new OrderDetailResponse(
                     b.OrderDetailId,
                     b.MealId,
                     b.Meal.MealName,
-                    b.Meal.Price,
+                    (int)b.Meal.Price,
                     b.Meal.ImageUrl,
                     b.Quantity,
-                    b.UnitPrice
+                    (int)b.UnitPrice
                 )).ToArray())).AsQueryable();
